# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ShareIT API",
    "version": "1.0.0",
    "description": "API documentation for ShareIT file sharing service"
  },
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login to get access token",
        "description": "Authenticate user and receive JWT token and API key",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "User's username"
                  },
                  "password": {
                    "type": "string",
                    "description": "User's password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "JWT token for authentication"
                    },
                    "apiKey": {
                      "type": "string",
                      "description": "API key for file operations"
                    },
                    "expiresIn": {
                      "type": "integer",
                      "description": "Token expiration time in milliseconds"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "username": {
                          "type": "string",
                          "description": "User's username"
                        },
                        "email": {
                          "type": "string",
                          "description": "User's email address"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Invalid credentials"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
      "FileInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "File ID"
          },
          "originalName": {
            "type": "string",
            "description": "Original file name"
          },
          "size": {
            "type": "integer",
            "description": "File size in bytes"
          },
          "mimeType": {
            "type": "string",
            "description": "File MIME type"
          }
        }
      }
    }
  }
} 